Refactors Action, adds Moves Action now acts as a controller to pass everything through to the relevant Move. Moves return a result which can be tweeted, and the moves themselves are responsible for updating models.
TODO: the way we identify a fight should change. it should be identifiable by the two usernames combined, NOT searching for a specific challenger and challenged. this data should still be kept, but we need a way to pull up the correct fight regardless of the order in which usernames are tweeted.